data.py file restructure
Remove unnecessary import
