+++ src/eatery/__init__.py
@@ -1,0 +1,5 @@
+from .campus_eatery import parse_eatery
+
+from .collegetown_eatery import parse_collegetown_eateries
+
+from .static_eatery import parse_static_eateries
